driver_custom.py > driver_extension.py get rid of get_by_type function change type hinting for by_type
